Despues del 2o dia de laboratorio casi funciona el objeto PUERTO.
Hay un problema al negar los bits de fila en la funcion fila_a_puerto donde ! no niega bit a bit sino que jode todo...

+++ src/hardware.c
@@ -67,7 +67,7 @@ char tecla_pulsada(Puerto *puerto)
                 {
                     while (lee_puertoE() & fila_mask); //Esperamos a que se suelte
                     retardo(RET_REBOTES); //Retardo antirrebotes
-                    return teclas[fila][columna]; //Devolvemos la tecla pulsada
+                    return teclas[(int)fila][(int)columna]; //Devolvemos la tecla pulsada
                 }
             }//Siguiente columna
         }// Exploración finalizada sin encontrar una tecla pulsada
